pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "earnings-poc-btg"
version = "0.1.0"
description = "Pipeline de extração de dados financeiros de earnings releases usando LangGraph + LLMs"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "openai>=1.0",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "black>=24.0",
    "ruff>=0.4",
]
ocr = [
    "unstructured[pdf]>=0.15",
]

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
